feat: Add SBS 3D video mode
Adds a ShaderGraphMaterial that splits the texture in half and displays the halves in stereo. The shader was pulled from [KhaosT's example](https://github.com/KhaosT/RealityKitSideBySideRenderExample) though that uses a metallic surface which messes up the colors. This works pretty flawlessly with half-width side-by-side video, performing better than Moon Video playing the same [example video](https://vimeo.com/167717612). It can probably be extended to handle full-width side-by-side video which would provide much better quality.

I've tested it with quite a bit of Half-Life 2 + Reshade, as well as Rise of the Tomb Raider's built in side-by-side 3D. All of the caveats of half-width side-by-side 3D apply: you're losing half the resolution. I only added this to the  not attempt to add this to `RealityKitStreamView` because I assumed it couldn't use the ShaderGraphMaterial, but I didn't actually check.